requirements.txt
flask
pymupdf
google-generativeai
fpdf
unidecode
python-docx